gcc12: fix bad array subscript on blob_start
Avoid "array subscript partly outside array bounds" due to
blob_memcpy doing 32-bit writes when blob_start was defined
uint8_t.

+++ src/reboot.c
@@ -143,8 +143,8 @@ asm(
 );
 
 /** Include the relocatable shim code */
-extern uint8_t blob_start;
-extern uint8_t blob_end;
+extern uint32_t blob_start;
+extern uint32_t blob_end;
 
 static void busy_wait(int n)
 {
